requirements.txt
aiofiles==23.1.0
alembic==1.11.1
anyio==3.7.0
asyncio==3.4.3
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.96.0
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.24.3
pandas==2.0.2
passlib==1.7.4
pycparser==2.21
pydantic==1.10.9
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
